+++ Mosquitto/lib/util_mosq.c
@@ -184,8 +184,8 @@ int mosquitto_topic_matches_sub(const char *sub, const char *topic, bool *result
 
 	if(!sub || !topic || !result) return MOSQ_ERR_INVAL;
 
-	slen = (int)strlen(sub);
-	tlen = (int)strlen(topic);
+	slen = strlen(sub);
+	tlen = strlen(topic);
 
 	if(slen && tlen){
 		if((sub[0] == '$' && topic[0] != '$')
